requirements.txt
google-api-python-client
google-auth
pydantic
python-dotenv
google-adk